Add rule to disallow `this.$` to prepare apps to remove jQuery
In modern browsers jQuery is less and less needed as many features are not built in
into modern browsers and the bugs that jQuery used to help you avoid have been fixed.
However, jQuery still adds ~30KB (min + gzip) to every Ember app, which can add up to
1 second of download + parse times on mobile devices. Is is not possible to remove
jQuery from Ember apps, and one of the main features in Ember that cannot be used on
an app that intends to remove jQuery in a near future is `this.$`, either inside
component's methods or within tests.

This PR adds a rule, that is entirely options and **I do not think** it should
be enabled by default to yell at anyone who uses `this.$`.

Please let me know what else has to be done (documentation, etc...) to make
this mergeable.