Prevent error when destructured path is not in known globals.
Prior to these changes, using destructuring that resulted in a path that
was not known to ember-rfc176-data's `globals.json` would throw an error
(because we eagerly accessed `match[1]`).

This adds a test for the failure scenario and ensures that only usage of
global paths that are listed in `globals.json` are reported as errors by
`new-module-imports`.