test: add null output assertions for lint rules / test cases with no autofixer.
If there's no autofixer for a particular rule or test case, we should assert that there is no output (`output` refers to the fixed version of the code produced by the autofixer). This helps make our tests more comprehensive, especially so that the tests will catch any future autofixer changes.

The documentation for `output` is here: https://eslint.org/docs/developer-guide/nodejs-api#ruletester