feat: add `ignoreNestedPaths` option (default true) to `no-get` rule
Having this rule ignore nested paths like `this.get('some.nested.property')` (which is the default behavior) can be helpful while adopting this rule to avoid breaking existing code. But long-term, developers may want to entirely get rid of `get()` usage, and can use this new option to do so.