refactor: avoid using file name in rules
The original logic for determining if a Identifier was of a specific Ember class used the file name to make this decision, rather than resolving the import of the Identifier back to the Ember module it was imported from.

This ended up causing a number of issues in lint rule in cases where people places a class in a location that was unexpected based on Ember’s convention. Such issues include #601 and #430.

Making this change avoids the need for something like #213, which adds additional logic based on file names to prevent some of these issues.

Closes #590