docs: mention if a rule is fixable in its documentation
Before this change, we only indicated which rules are fixable in the top-level README, but not in individual rule docs.

This change adds a message to individual rule docs that matches the convention of eslint core.

Also adds a test to ensure that the correct rule docs include this notice.