feat: add autofixer to `no-get-with-default` rule
The [no-get-with-default](https://github.com/ember-cli/eslint-plugin-ember/blob/master/docs/rules/no-get-with-default.md) rule is now a recommended rule in the v8 release so it's even more useful to have an autofixer for it.

Before:
```js
const test = this.getWithDefault('property', []);
```
After:
```js
const test = (this.get('property') === undefined ? [] : this.get('property'));
```

This rule converts to `this.get('property')` for safety. The `no-get` rule can then convert it to ES5 getters (`this.property`) if safe.