docs: switch from redundant dynamic error message to static error message in no-get rule
Maintaining a dynamic error message showing the before and after code has become a burden to maintain given the many variations / options / edge cases that the `no-get` has grown to support.

It's also redundant to show the before and after code, given that the user already knows what the before code is, and the rule can autofix to the recommended code.