Disallow use of unclosed attribute selector blocks in test helpers
Tests which use test-helpers that rely on querySelector or querySelectorAll to
lookup an element in the dom are able to use invlid attribute selectors which do not have a closing bracket.
This happens to work today due to a quirk in the CSS selector spec and is marked as WontFix by Chromium.
We should be preventing invalid selectors from being used in test-helpers as they are likely to stop working
at some point and are not valid selector syntax.

fixes #979