refactor: improve detection of global window methods in prefer-ember-test-helpers rule
Eliminates all of the custom logic in the rule for detecting if the `blur` / `find` / `focus` window functions are defined locally and replaces it with the ReferenceTracker class from eslint-utils.

There was no specific known bug with the rule before, but this refactor will make the rule more robust in its ability to detect calls to the global window functions that we are trying to disallow.

https://eslint-utils.mysticatea.dev/api/scope-utils.html#referencetracker-class