Fix an eslint assertion within jquery-ember-run
The included test demonstrates a case that would trigger an assertion
within eslint. It's not correct to try to `report(expression.callee)`
without asserting that the expression is a CallExpression. Lots of other
kinds of nodes can appear here instead, and the AssignmentExpression in
my example is only one of the possibilities.

The refactoring into two helper functions here was forced by the
complexity eslint rule this repo uses.