Remove `no-array-prototype-extensions` rule from `recommended` config (#1562)
* Remove no-array-prototype-extensions from recommended

The `no-array-prototype-extensions` rule is trying to do an impossible thing. It's going to produce a never-ending stream of false positives. You cannot statically tell which calls are actually relying on the array prototype extensions.

A correct implementation is probably possible, but only in typescript (where eslint rules can take advantage of type information), not in javascript.

I'm all in favor of pushing people away from using array prototype extensions, but that is going to require runtime implementation. It cannot be checked statically. A rule that is so often wrong will just encourage people not to trust it anyway. I don't think it should be in the recommended set.

Examples:

https://github.com/ember-cli/eslint-plugin-ember/issues/1561
https://github.com/ember-cli/eslint-plugin-ember/pull/1552
https://github.com/ember-cli/eslint-plugin-ember/pull/1547
https://github.com/ember-cli/eslint-plugin-ember/pull/1546
https://github.com/ember-cli/eslint-plugin-ember/pull/1544
https://github.com/ember-cli/eslint-plugin-ember/pull/1543
https://github.com/ember-cli/eslint-plugin-ember/pull/1538
https://github.com/ember-cli/eslint-plugin-ember/pull/1539
https://github.com/ember-cli/eslint-plugin-ember/pull/1536

* docs: mentioned recommend rule removal in changelog

* docs: explain why no-array-prototype-extensions is not in recommended config

Co-authored-by: Bryan Mishkin <698306+bmish@users.noreply.github.com>